version 4 - after package refactoring - separate undertow connection low-level classes to separate package
