+++ client/src/main/java/com/networknt/client/simplepool/undertow/SimpleSSLContextMaker.java
@@ -1,4 +1,4 @@
-package com.networknt.client.http;
+package com.networknt.client.simplepool.undertow;
 
 import com.networknt.client.ClientConfig;
 import com.networknt.client.ssl.ClientX509ExtendedTrustManager;
